requirements.txt
torch==1.8.1
torch_geometric==1.7.1
tensorboard==2.5.0
uproot3==3.14.4
pandas==1.2.4
numpy==1.20.2
tqdm==4.60.0
scikit-learn==0.24.1
pyyaml==5.4.1
matplotlib==3.4.1